requirements.txt
python-telegram-bot>=21.0
httpx==0.27.2
reportlab>=4.2.0
python-dotenv==1.0.1
google-genai>=1.0.0
pydantic>=2.0.0
requests
fastapi
uvicorn
